init in-progress
