--- .eslintrc.json
@@ -1,84 +1,0 @@
-{
-    "env": {
-        "browser": true,
-        "commonjs": true,
-        "es6": true,
-        "node": true,
-        "mocha": true
-    },
-    "parser": "babel-eslint",
-    "extends": "airbnb",
-    "parserOptions": {
-        "ecmaFeatures": {
-            "jsx": true
-        },
-        "sourceType": "module"
-    },
-    "plugins": [
-        "import"
-    ],
-    "rules": {
-        "no-multi-spaces": 0,
-        "key-spacing": 0,
-        "indent": [
-            "error",
-            4,
-            {
-                "SwitchCase": 1
-            }
-        ],
-        "no-console": [
-            "error",
-            {
-                "allow": [
-                    "warn",
-                    "error",
-                    "info"
-                ]
-            }
-        ],
-        "no-underscore-dangle": 0,
-        "comma-dangle": 0,
-        "no-useless-constructor": 0,
-        "max-len": 0,
-        "import/no-unresolved": 0,
-        "import/extensions": 0,
-        "import/no-extraneous-dependencies": 0,
-        "func-names": 0,
-        "prefer-arrow-callback": 0,
-        "no-const-assign": "warn",
-        "no-this-before-super": "warn",
-        "no-undef": "warn",
-        "no-unreachable": "warn",
-        "no-unused-vars": "warn",
-        "constructor-super": "warn",
-        "valid-typeof": "warn",
-        "react/jsx-equals-spacing": 0,
-        "react/jsx-filename-extension": 0,
-        "react/jsx-indent-props": 0,
-        "react/prop-types": 0,
-        "react/jsx-indent": [
-            "error",
-            4
-        ],
-        "react/forbid-prop-types": 0,
-        "react/require-default-props": 0,
-        "jsx-a11y/href-no-hash": 0,
-        "arrow-parens": 0,
-        "import/prefer-default-export": 0,
-        "class-methods-use-this": 0
-    },
-    "globals": {
-        "expect": true,
-        "assert": true,
-        "should": true,
-        "window": true,
-        "document": true,
-        "HTMLElement": true,
-        "CustomEvent": true,
-        "Event": true,
-        "fetch": true,
-        "MutationObserver": true,
-        "sinon": true
-    }
-}
