--- build-config/loaders.js
@@ -1,6 +1,0 @@
-module.exports = [
-    {
-        test: /\.(js)$/,
-        use: { loader: 'babel-loader' }
-    }
-];
