+++ build-config/prettier/index.js
@@ -1,0 +1,6 @@
+module.exports = {
+    printWidth: 120,
+    tabWidth: 4,
+    singleQuote: true,
+    arrowParens: 'always'
+};
