+++ build-config/tslint/default.json
@@ -1,0 +1,19 @@
+{
+    "extends": ["tslint:recommended", "tslint-config-airbnb", "tslint-config-prettier"],
+    "rulesDirectory": ["tslint-plugin-prettier"],
+    "rules": {
+        "prettier": true,
+        "interface-name": false,
+        "no-this-assignment": [true, {
+            "allowed-names": ["^self$"],
+            "allow-destructuring": true
+        }],
+        "object-literal-sort-keys": false,
+        "no-console": [true, "log"],
+        "indent": [true, "spaces", 4],
+        "member-access": false
+    },
+    "jsRules": {
+        "object-literal-sort-keys": false
+    }
+}
