+++ build-config/webpack/loaders/index.js
@@ -1,0 +1,14 @@
+const path = require('path');
+
+module.exports = () =>  [
+    {
+        exclude: /node_modules/,
+        loader: 'babel-loader',
+        test: /\.(tsx?)|.(jsx?)$/
+    },
+    {
+        enforce: 'pre',
+        test: /\.js$/,
+        loader: 'source-map-loader'
+    }
+];
