--- jsdoc.json
@@ -1,24 +1,0 @@
-{
-    "tags": {
-        "allowUnknownTags": false
-    },
-    "source": {
-        "include": ["src"],
-        "exclude": ["node_modules","docs", "dist"],
-        "includePattern": ".js$"
-    },
-    "plugins": [
-        "plugins/markdown"
-    ],
-    "opts": {
-        "template": "node_modules/minami",
-        "encoding": "utf8",
-        "destination": "docs/",
-        "recurse": true,
-        "verbose": true
-    },
-    "templates": {
-        "cleverLinks": false,
-        "monospaceLinks": false
-    }
-}
