+++ package.json
@@ -1,72 +1,73 @@
 {
-  "name": "baseui-wc-base-component",
-  "version": "0.4.6",
-  "description": "re-usable base customElement to avoid repeated bootstrap code, flexible to choose own rendering or use hyperHTML/LitHTML.",
-  "main": "dist/baseui-wc-base-component.js",
-  "module": "src/index.js",
-  "scripts": {
-    "start": "npm i && npm run server",
-    "clean": "test -d 'dist' && rm -r dist/* || echo 'Build folder is clean!'",
-    "docs": "./node_modules/.bin/jsdoc -c jsdoc.json",
-    "lint": "./node_modules/.bin/eslint --fix  ./src/**/*.js --quiet",
-    "build": "npm run clean && npm run build-dev && npm run build-min",
-    "build-dev": "NODE_ENV=production ./node_modules/.bin/webpack --config webpack.prod.js",
-    "build-min": "NODE_ENV=production ./node_modules/.bin/webpack --env.min --config webpack.prod.js",
-    "server": "export BABEL_ENV=production && ./node_modules/.bin/webpack-dev-server --config webpack.dev.js",
-    "bundleServ": "./node_modules/.bin/live-server --port=1919 --host=0.0.0.0 --open='./demo/'",
-    "release": "npm run lint && npm run build && npm run docs",
-    "test": "NODE_ENV=test jest --config jest.config.js",
-    "test:watch": "NODE_ENV=test jest --watch"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/dsadhanala/baseui-wc-base-component.git"
-  },
-  "keywords": [
-    "web component",
-    "customElement",
-    "baseUI customElement",
-    "customElement with lit-html",
-    "customElement with hyper-html",
-    "ES6 template literals",
-    "ES6 mixin"
-  ],
-  "author": "dsadhanala@gmail.com",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/dsadhanala/wc-base-component/issues"
-  },
-  "homepage": "https://github.com/dsadhanala/wc-base-component#readme",
-  "dependencies": {
-    "hyperhtml": "^2.10.1",
-    "lit-html": "^0.10.0"
-  },
-  "devDependencies": {
-    "babel-core": "^6.26.3",
-    "babel-eslint": "^8.2.3",
-    "babel-jest": "^22.4.3",
-    "babel-loader": "^7.1.4",
-    "babel-plugin-transform-builtin-classes": "^0.6.1",
-    "babel-plugin-transform-object-rest-spread": "^6.26.0",
-    "babel-polyfill": "^6.26.0",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-stage-0": "^6.24.1",
-    "basichtml": "^0.16.0",
-    "eslint": "^4.19.1",
-    "eslint-config-airbnb": "^16.1.0",
-    "eslint-plugin-import": "^2.11.0",
-    "eslint-plugin-jsx-a11y": "^6.0.3",
-    "eslint-plugin-react": "^7.8.2",
-    "jest": "^22.4.3",
-    "jest-environment-node": "^22.4.3",
-    "jsdoc": "^3.5.5",
-    "live-server": "^1.2.0",
-    "minami": "^1.2.3",
-    "prettier-eslint": "^8.8.1",
-    "sizzle": "^2.3.3",
-    "webpack": "^4.8.3",
-    "webpack-cli": "^2.1.3",
-    "webpack-dev-server": "^3.1.4",
-    "webpack-merge": "^4.1.2"
-  }
+    "name": "baseui-wc-base-component",
+    "version": "0.5.0",
+    "description": "re-usable base customElement to avoid repeated bootstrap code, flexible to choose own rendering or use hyperHTML/LitHTML.",
+    "main": "dist/baseui-wc-base-component.js",
+    "module": "src/index.js",
+    "scripts": {
+        "start": "npm i && npm run server",
+        "tc": "npx tsc --noEmit",
+        "tcw": "npm run type-check -- -w",
+        "build:types": "npx tsc --emitDeclarationOnly",
+        "lint": "npx tslint './src/**/*.{js,jsx,ts,tsx}' --project tsconfig.json",
+        "lint:fix": "npx tslint './src/**/*.{js,jsx,ts,tsx}' --fix --project tsconfig.json",
+        "server": "NODE_ENV=development npx webpack-dev-server --mode development --config webpack.dev.js",
+        "build": "npm run build-dev && npm run build-min",
+        "build-dev": "NODE_ENV=production webpack --config webpack.prod.js",
+        "build-min": "NODE_ENV=production webpack --env.min --config webpack.prod.js",
+        "release": "npm run lint:fix && npm run build:types && npm run build",
+        "deploy": "npx https://gist.github.com/dsadhanala/0157697871892caaf2793170a61e6ebd",
+        "bundleServ": "./node_modules/.bin/live-server --port=1919 --host=0.0.0.0 --open='./demo/'"
+    },
+    "repository": {
+        "type": "git",
+        "url": "git+https://github.com/dsadhanala/baseui-wc-base-component.git"
+    },
+    "keywords": [
+        "web component",
+        "customElement",
+        "baseUI customElement",
+        "customElement with lit-html",
+        "customElement with hyper-html",
+        "ES6 template literals",
+        "ES6 mixin"
+    ],
+    "author": "dsadhanala@gmail.com",
+    "license": "MIT",
+    "bugs": {
+        "url": "https://github.com/dsadhanala/wc-base-component/issues"
+    },
+    "homepage": "https://github.com/dsadhanala/wc-base-component#readme",
+    "dependencies": {
+        "hyperhtml": "^2.14.0",
+        "lit-html": "^0.12.0"
+    },
+    "devDependencies": {
+        "@babel/cli": "^7.1.2",
+        "@babel/core": "^7.1.2",
+        "@babel/plugin-proposal-class-properties": "^7.1.0",
+        "@babel/plugin-proposal-decorators": "^7.1.2",
+        "@babel/plugin-proposal-object-rest-spread": "^7.0.0",
+        "@babel/polyfill": "^7.0.0",
+        "@babel/preset-env": "^7.1.0",
+        "@babel/preset-typescript": "^7.1.0",
+        "babel-loader": "^8.0.4",
+        "basichtml": "^0.19.0",
+        "clean-webpack-plugin": "^0.1.19",
+        "fork-ts-checker-webpack-plugin": "^0.4.10",
+        "husky": "^1.1.2",
+        "live-server": "^1.2.0",
+        "prettier": "^1.14.3",
+        "sizzle": "^2.3.3",
+        "tslint": "^5.11.0",
+        "tslint-config-airbnb": "^5.11.0",
+        "tslint-config-prettier": "^1.15.0",
+        "tslint-config-standard": "^8.0.1",
+        "tslint-plugin-prettier": "^2.0.0",
+        "typescript": "^3.1.3",
+        "webpack": "^4.8.3",
+        "webpack-cli": "^3.1.2",
+        "webpack-dev-server": "^3.1.4",
+        "webpack-merge": "^4.1.2"
+    }
 }
