+++ src/_global-types/index.ts
@@ -1,0 +1,1 @@
+//
