--- src/base-component/_tests/index.js
@@ -1,14 +1,0 @@
-import BaseCustomElement from '../index.js';
-// const { init, HTMLElement } = require('basichtml');
-
-// init();
-// global.HTMLElement = HTMLElement;
-
-// describe('base element', () => {
-//
-// });
-test('adds 1 + 2 to equal 3', () => {
-    // const element = new BaseCustomElement();
-    // expect(element).toBe();
-    console.log('hey');
-});
