+++ src/throw-error/index.ts
