+++ tsconfig.json
@@ -1,0 +1,25 @@
+{
+    "compilerOptions": {
+        "pretty": true,
+        "moduleResolution": "node",
+        "target": "esnext",
+        "module": "umd",
+        "skipLibCheck": true,
+        "lib": ["es2015", "es2016", "es2017", "dom"],
+        "strict": true,
+        "sourceMap": true,
+        "declaration": true,
+        // "allowJs": true,
+        "allowSyntheticDefaultImports": true,
+        "esModuleInterop": true,
+        "experimentalDecorators": true,
+        "emitDecoratorMetadata": true,
+        "declarationDir": "dist/types",
+        "outDir": "@types",
+        "rootDir": ".",
+        "typeRoots": ["./node_modules/@types", "./src/global-types"]
+    },
+    "include": ["src", "demo/tests"],
+    "exclude": ["node_modules"],
+    "files": ["./src/index.ts", "./demo/tests/index.ts"]
+}
